Tests/LibWeb/Text/expected/multi-click-user-select-none-element.txt

double-click selected: ""
triple-click selected: ""
